server/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
click==8.1.7
fastapi==0.103.0
h11==0.14.0
httptools==0.6.0
idna==3.4
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
